+++ pkg/lightning/tikv/local_sst_writer.go
@@ -19,10 +19,16 @@ import (
 	"encoding/binary"
 
 	"github.com/google/uuid"
+	rocks "github.com/lance6716/pebble"
+	rocksbloom "github.com/lance6716/pebble/bloom"
+	"github.com/lance6716/pebble/objstorage/objstorageprovider"
+	rockssst "github.com/lance6716/pebble/sstable"
+	"github.com/lance6716/pebble/vfs"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/kvproto/pkg/import_sstpb"
 	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/pingcap/tidb/pkg/util/codec"
+	"github.com/pingcap/tidb/pkg/util/intest"
 	pd "github.com/tikv/pd/client"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
@@ -130,48 +136,69 @@ func write2ImportService4Test(
 	return resp.Metas, nil
 }
 
+// writeCFWriter generates SST files for TiKV's write column family.
+type writeCFWriter struct {
+	sstWriter *rockssst.Writer
+	ts        uint64
+}
+
+// newWriteCFWriter creates a new writeCFWriter. Currently `identity` is acquired
+// from a real SST file generated by TiKV.
+func newWriteCFWriter(
+	sstPath string,
+	ts uint64,
+	identity *rockssst.Identity,
+) (*writeCFWriter, error) {
+	f, err := vfs.Default.Create(sstPath)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	writable := objstorageprovider.NewFileWritable(f)
+	writer := rockssst.NewWriter(writable, rockssst.WriterOptions{
+		// TODO(lance6716): should read TiKV config to know compression algorithm.
+		Compression:  rocks.ZstdCompression,
+		FilterPolicy: rocksbloom.FilterPolicy(10),
+		MergerName:   "nullptr",
+		TablePropertyCollectors: []func() rockssst.TablePropertyCollector{
+			func() rockssst.TablePropertyCollector {
+				return newMVCCPropCollector(ts)
+			},
+			func() rockssst.TablePropertyCollector {
+				return newRangePropertiesCollector()
+			},
+			// TODO(lance6716): check if we should trigger titan
+			func() rockssst.TablePropertyCollector {
+				return mockCollector{name: "BlobFileSizeCollector"}
+			},
+		},
+	}, identity)
+	return &writeCFWriter{sstWriter: writer, ts: ts}, nil
+}
+
+// set mimic TiKV's `TxnSstWriter.put`. ref
+// https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/sst_importer/src/sst_writer.rs#L92.
+func (w *writeCFWriter) set(key, value []byte) error {
+	intest.Assert(isShortValue(value), "not implemented, need to write to default CF")
+
+	actualKey := make([]byte, 0, 1+len(key)+8)
+	actualKey = append(actualKey, 'z')
+	actualKey = codec.EncodeBytes(actualKey, key)
+	actualKey = binary.BigEndian.AppendUint64(actualKey, ^w.ts)
+
+	actualValue := make([]byte, 0, 1+binary.MaxVarintLen64+1+1+len(value))
+	actualValue = append(actualValue, 'P')
+	actualValue = binary.AppendUvarint(actualValue, w.ts)
+	actualValue = append(actualValue, 'v')
+	actualValue = append(actualValue, byte(len(value)))
+	actualValue = append(actualValue, value...)
+
+	return errors.Trace(w.sstWriter.Set(actualKey, actualValue))
+}
+
+func (w *writeCFWriter) close() error {
+	return errors.Trace(w.sstWriter.Close())
+}
+
 func isShortValue(val []byte) bool {
 	return len(val) <= 255
-}
-
-func encodeWriteBatch2KV(batch *import_sstpb.WriteBatch) [][2][]byte {
-	ts := batch.CommitTs
-	kvs := make([][2][]byte, 0, len(batch.Pairs))
-	for _, pair := range batch.Pairs {
-		keyWithTS := []byte{'z'}
-		keyWithTS = codec.EncodeBytes(keyWithTS, pair.Key)
-		keyWithTS = binary.BigEndian.AppendUint64(keyWithTS, ^ts)
-		if isShortValue(pair.Value) {
-			// put flag
-			val := append([]byte{}, 'P')
-			val = binary.AppendUvarint(val, ts)
-			// short value flag
-			val = append(val, 'v')
-			val = append(val, byte(len(pair.Value)))
-			val = append(val, pair.Value...)
-
-			kvs = append(kvs, [2][]byte{keyWithTS, val})
-		} else {
-			panic("not implemented, need to write to default CF")
-		}
-	}
-	return kvs
-}
-
-func encodeKVs4Test(
-	sortedKVs [][2][]byte,
-	ts uint64,
-) [][2][]byte {
-	batch := &import_sstpb.WriteBatch{
-		CommitTs: ts,
-		Pairs:    make([]*import_sstpb.Pair, 0, len(sortedKVs)),
-	}
-	for _, kv := range sortedKVs {
-		batch.Pairs = append(batch.Pairs, &import_sstpb.Pair{
-			Key:   kv[0],
-			Value: kv[1],
-		})
-	}
-
-	return encodeWriteBatch2KV(batch)
 }
